Refactor plotly plots into their own function
