+++ src/gemdat/plots/plotly/_jumps_vs_distance.py
@@ -1,0 +1,75 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+import numpy as np
+import pandas as pd
+import plotly.express as px
+import plotly.graph_objects as go
+
+if TYPE_CHECKING:
+    from gemdat import Jumps
+
+
+def jumps_vs_distance(*,
+                      jumps: Jumps,
+                      jump_res: float = 0.1,
+                      n_parts: int = 1) -> go.Figure:
+    """Plot jumps vs. distance histogram.
+
+    Parameters
+    ----------
+    jumps : Jumps
+        Input jumps data
+    jump_res : float, optional
+        Resolution of the bins in Angstrom
+    n_parts : int
+        Number of parts for error analysis
+
+    Returns
+    -------
+    fig : plotly.graph_objects.Figure
+        Output figure
+    """
+    sites = jumps.sites
+    trajectory = jumps.trajectory
+    lattice = trajectory.get_lattice()
+
+    pdist = lattice.get_all_distances(sites.frac_coords, sites.frac_coords)
+
+    bin_max = (1 + pdist.max() // jump_res) * jump_res
+    n_bins = int(bin_max / jump_res) + 1
+    x = np.linspace(0, bin_max, n_bins)
+
+    bin_idx = np.digitize(pdist, bins=x)
+    data = []
+    for transitions_part in jumps.split(n_parts=n_parts):
+        counts = np.zeros_like(x)
+        for idx, n in zip(bin_idx.flatten(),
+                          transitions_part.matrix().flatten()):
+            counts[idx] += n
+        for idx in range(n_bins):
+            if counts[idx] > 0:
+                data.append((x[idx], counts[idx]))
+
+    df = pd.DataFrame(data=data, columns=['Displacement', 'count'])
+
+    grouped = df.groupby(['Displacement'])
+    mean = grouped.mean().reset_index().rename(columns={'count': 'mean'})
+    std = grouped.std().reset_index().rename(columns={'count': 'std'})
+    df = mean.merge(std, how='inner')
+
+    if n_parts == 1:
+        fig = px.bar(df, x='Displacement', y='mean', barmode='stack')
+    else:
+        fig = px.bar(df,
+                     x='Displacement',
+                     y='mean',
+                     error_y='std',
+                     barmode='stack')
+
+    fig.update_layout(title='Jumps vs. Distance',
+                      xaxis_title='Distance (Å)',
+                      yaxis_title='Number of jumps')
+
+    return fig
